requirements.txt
# Core
streamlit
streamlit-quill

# Streamlit components
streamlit-ace
streamlit-discourse
streamlit-disqus
streamlit-elements
streamlit-pandas-profiling
streamlit-player

# Dependencies
pandas-profiling
requests

openai
pandasai
astor


